Unify the EL7 and EL8 content and fix the SELinux policy packaging
The actual SELinux policy content was identical between EL7 and EL8,
with minor differences in the RPM spec file. Furthermore, the scripts
for actually doing these builds were not significantly different.

This unifies them so that we don't suffer maintaining two versions
for no reason.

Additionally, rewrite the SELinux policy package RPM spec file to align
with SELinux module packaging to make it easier to ship in distributions.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>